refactor: apply path alias
