+++ apps/server/src/app.module.ts
@@ -1,15 +1,14 @@
 import { Module } from '@nestjs/common';
 
-import { AuthModule } from './auth/auth.module';
-import { PrismaModule } from './prisma/prisma.module';
-import { PrismaService } from './prisma/prisma.service';
-import { QuestionsModule } from './questions/questions.module';
-import { RepliesModule } from './replies/replies.module';
-import { SessionsModule } from './sessions/sessions.module';
-import { SessionsAuthModule } from './sessions-auth/sessions-auth.module';
-import { UsersModule } from './users/users.module';
-
-import { UploadModule } from '@src/upload/upload.module';
+import { AuthModule } from '@auth/auth.module';
+import { PrismaModule } from '@prisma-alias/prisma.module';
+import { PrismaService } from '@prisma-alias/prisma.service';
+import { QuestionsModule } from '@questions/questions.module';
+import { RepliesModule } from '@replies/replies.module';
+import { SessionsModule } from '@sessions/sessions.module';
+import { SessionsAuthModule } from '@sessions-auth/sessions-auth.module';
+import { UploadModule } from '@upload/upload.module';
+import { UsersModule } from '@users/users.module';
 @Module({
   imports: [
     UsersModule,
