+++ apps/server/src/auth/auth.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
+import { UsersRepository } from '@users/users.repository';
 
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { JwtAuthGuard } from './jwt-auth.guard';
-import { PrismaModule } from '../prisma/prisma.module';
-import { UsersRepository } from '../users/users.repository';
+
+import { PrismaModule } from '@prisma-alias/prisma.module';
 
 @Module({
   imports: [JwtModule.register({}), PrismaModule],
