+++ apps/server/src/questions/questions.module.ts
@@ -1,4 +1,4 @@
-import { Module } from '@nestjs/common';
+import x{ Module } from '@nestjs/common';
 
 import { QuestionsController } from './questions.controller';
 import { QuestionsRepository } from './questions.repository';
