+++ apps/server/src/replies/dto/toggle-reply-like.dto.ts
@@ -1,3 +1,3 @@
-import { BaseDto } from '@src/common/base.dto';
+import { BaseDto } from '@common/base.dto';
 
 export class ToggleReplyLikeDto extends BaseDto {}
