+++ apps/server/src/sessions-auth/sessions-auth.controller.ts
@@ -4,8 +4,9 @@ import { ApiTags } from '@nestjs/swagger';
 import { SessionAuthDto } from './dto/session-auth.dto';
 import { SessionsAuthService } from './sessions-auth.service';
 import { AuthSessionsSwagger } from './swagger/sessions-auth.swagger';
-import { JwtPayloadInterceptor } from '../common/interceptors/jwt-payload.interceptor';
-import { TransformInterceptor } from '../common/interceptors/transform.interceptor';
+
+import { JwtPayloadInterceptor } from '@common/interceptors/jwt-payload.interceptor';
+import { TransformInterceptor } from '@common/interceptors/transform.interceptor';
 @ApiTags('session-auth')
 @UseInterceptors(TransformInterceptor)
 @UseInterceptors(JwtPayloadInterceptor)
