+++ apps/server/src/sessions-auth/sessions-auth.module.ts
@@ -4,7 +4,8 @@ import { JwtModule } from '@nestjs/jwt';
 import { SessionsAuthController } from './sessions-auth.controller';
 import { SessionsAuthRepository } from './sessions-auth.repository';
 import { SessionsAuthService } from './sessions-auth.service';
-import { PrismaModule } from '../prisma/prisma.module';
+
+import { PrismaModule } from '@prisma-alias/prisma.module';
 
 @Module({
   imports: [JwtModule.register({}), PrismaModule],
