+++ apps/server/src/users/users.module.ts
@@ -3,7 +3,8 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersRepository } from './users.repository';
 import { UsersService } from './users.service';
-import { PrismaModule } from '../prisma/prisma.module';
+
+import { PrismaModule } from '@prisma-alias/prisma.module';
 
 @Module({
   imports: [PrismaModule],
