+++ apps/server/src/users/users.repository.ts
@@ -1,11 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
 
-import { DatabaseException } from '../common/exceptions/resource.exception';
-import { PRISMA_ERROR_CODE } from '../prisma/prisma.error';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UserConflictException } from './exceptions/user.exception';
+
+import { DatabaseException } from '@common/exceptions/resource.exception';
+import { PRISMA_ERROR_CODE } from '@prisma-alias/prisma.error';
 
 @Injectable()
 export class UsersRepository {
